pyproject.toml
[project]
name = "tcm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.0.1",
    "langchain-milvus>=0.2.2",
    "langchain-openai>=1.0.2",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "torch",
]


[tool.uv]
index-url = "https://mirrors.ustc.edu.cn/pypi/simple"
extra-index-url = ["https://mirrors.ustc.edu.cn/pypi/simple/"]

[tool.uv.sources]
torch = { path = "../../gpu_pkg/torch-2.8.0+cu129-cp313-cp313-win_amd64.whl" }
